backend/pyproject.toml
[tool.poetry]
name = "connect3"
version = "0.1.0"
description = ""
authors = ["kiril <kirill.301522@gmail.com>"]
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.3"
python-dotenv = "^1.0.1"
flake8 = "^7.0.0"
isort = "^5.13.2"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
factory-boy = "^3.3.0"
pillow = "^10.2.0"
djangorestframework = "^3.15.0"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
django-modeladmin-reorder = "^0.3.1"
setuptools = "^69.2.0"
qrcode = "^7.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
